Split on word boundaries for perfect decryption
Now words and the spaces between them are each separate blocks.
  E.g. "foo bar." -> "foo", " ", "bar", "."
Note that this does split up contractions -- "it's" is three blocks.

Reverted lower-casing. With the changes above, decryption perfectly
reproduces the input file, and down-casing would break that.
Case-insensitive search can go in version 2.

No need to pad the blocks in Enc, as it's done in TwoBlockEncrypt.